+++ app/models/observation.rb
@@ -58,12 +58,14 @@ class Observation < ApplicationRecord
   validates :publication_date, presence: true
   validates_presence_of :validation_status
 
+  before_save   :set_active_status
   after_create  :update_operator_scores
   after_destroy :update_operator_scores
   after_save    :update_operator_scores, if: 'publication_date_changed? || severity_id_changed?'
 
   include Activable
 
+  default_scope { where(is_active: true) }
 
   class << self
     def translated_types
@@ -87,5 +89,10 @@ class Observation < ApplicationRecord
 
   def update_operator_scores
     operator.calculate_observations_scores unless operator.nil?
+  end
+
+  def set_active_status
+    self.is_active = self.validation_status == 'Approved' ? true : false
+    nil
   end
 end